Resolve self-linting violations
Extract trailing_whitespace's candidate collection into a helper to
reduce cyclomatic complexity, wrap an overlong line, and move
indentation_width's helper visitors into their own file to respect
the file length limit.
